Refactor pipe components and move them into core library package
