+++ src/main/java/com/logistics/pipe/ChassisPipe.java
@@ -6,10 +6,10 @@ import com.logistics.core.lib.resource.ResourceId;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
 import com.logistics.pipe.item.ModuleItem;
 import com.logistics.core.lib.pipe.RandomTickModule;
-import com.logistics.pipe.modules.RoutingModule;
-import com.logistics.pipe.modules.TransferHandlerModule;
-import com.logistics.pipe.runtime.RoutePlan;
-import com.logistics.pipe.runtime.TravelingItem;
+import com.logistics.core.lib.pipe.RoutePlan;
+import com.logistics.core.lib.pipe.RoutingModule;
+import com.logistics.core.lib.pipe.TransferHandlerModule;
+import com.logistics.core.lib.pipe.TravelingItem;
 import net.minecraft.util.RandomSource;
 import com.logistics.pipe.ui.ChassisScreenHandler;
 import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
