+++ src/main/java/com/logistics/pipe/Pipe.java
@@ -12,10 +12,10 @@ import com.logistics.core.lib.pipe.DispatchableModule;
 import com.logistics.core.lib.pipe.Module;
 import com.logistics.core.lib.pipe.RandomTickModule;
 import com.logistics.core.lib.pipe.TickingModule;
-import com.logistics.pipe.modules.RoutingModule;
-import com.logistics.pipe.modules.TransferHandlerModule;
-import com.logistics.pipe.runtime.RoutePlan;
-import com.logistics.pipe.runtime.TravelingItem;
+import com.logistics.core.lib.pipe.RoutePlan;
+import com.logistics.core.lib.pipe.RoutingModule;
+import com.logistics.core.lib.pipe.TransferHandlerModule;
+import com.logistics.core.lib.pipe.TravelingItem;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
