+++ src/main/java/com/logistics/pipe/modules/ModSinkModule.java
@@ -1,5 +1,7 @@
 package com.logistics.pipe.modules;
 
+import com.logistics.core.lib.pipe.RoutingModule;
+
 import com.logistics.LogisticsPipe;
 import com.logistics.core.lib.pipe.*;
 import com.logistics.core.lib.pipe.Module;
@@ -8,9 +10,8 @@ import com.logistics.core.lib.storage.DirectionSerializer;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
 import com.logistics.core.lib.network.ILogisticsNetwork;
 import com.logistics.pipe.network.NetDbg;
-import com.logistics.pipe.network.NetworkRegistry;
-import com.logistics.pipe.runtime.RoutePlan;
-import com.logistics.pipe.runtime.TravelingItem;
+import com.logistics.core.lib.pipe.RoutePlan;
+import com.logistics.core.lib.pipe.TravelingItem;
 import com.logistics.pipe.ui.ModSinkScreenHandler;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -63,7 +64,7 @@ public class ModSinkModule implements Module, TickingModule, RoutingModule, Item
     @Override
     public void onDetach(PipeContext ctx) {
         if (!ctx.world().isClientSide()) {
-            ILogisticsNetwork network = NetworkRegistry.getNetwork(ctx.world(), ctx.pos());
+            ILogisticsNetwork network = ctx.network();
             if (network != null) network.unregisterSink(ctx.pos());
         }
     }
@@ -222,7 +223,7 @@ public class ModSinkModule implements Module, TickingModule, RoutingModule, Item
     }
 
     private void registerWithNetwork(PipeContext ctx) {
-        ILogisticsNetwork network = NetworkRegistry.getNetwork(ctx.world(), ctx.pos());
+        ILogisticsNetwork network = ctx.network();
         if (network != null) {
             network.registerSink(ctx.pos(), priority);
             network.registerGenericSinkInterest(ctx.pos());
