+++ src/test/java/com/logistics/pipe/runtime/TravelingItemTest.java
@@ -1,5 +1,7 @@
 package com.logistics.pipe.runtime;
 
+import com.logistics.core.lib.pipe.TravelingItem;
+
 import com.logistics.LogisticsPipe;
 import com.logistics.test.MinecraftTestEnvironment;
 import net.minecraft.core.Direction;
